+++ src/app/_hook/api/reviews/index.ts
@@ -1,0 +1,7 @@
+import { SortOption } from '@/app/_types/review.type'
+
+export const reviewKeys = {
+  all: ['review'] as const,
+  reviewList: (itemId: number, sortOption: SortOption) =>
+    [...reviewKeys.all, itemId, sortOption] as const,
+}
